requirements.txt
git+https://github.com/Rapptz/discord.py@5e6335750819588062790c6b93d604d3bc7c3862
pyyaml
requests
Pillow
imagehash
